requirements.txt
backcall==0.2.0
cycler==0.10.0
debugpy==1.4.1
decorator==5.0.9
entrypoints==0.3
greenlet==1.1.1
ipykernel==6.2.0
ipython==7.26.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==7.0.0
jupyter-core==4.7.1
kiwisolver
matplotlib
matplotlib-inline
nest-asyncio
numpy
pandas
parso
pexpect
pickleshare
Pillow
prompt-toolkit
psycopg2-binary
ptyprocess
Pygments
pyparsing
python-dateutil
pytz
pyzmq
scipy
seaborn
six
SQLAlchemy
tornado
traitlets
wcwidth
dash==2.6.0
dash-bootstrap-components
dash-bootstrap-templates
plotly